Refactor main.rs into modular structure
Split the monolithic 964-line main.rs into 9 focused modules:
- config.rs: Configuration structures and defaults (101 lines)
- stats.rs: Statistics tracking and health status (69 lines)
- notifier.rs: Email and Discord notification functionality (85 lines)
- utils.rs: Utility functions for logging, retries, and formatting (38 lines)
- webreg.rs: WebReg wrapper initialization and connection validation (34 lines)
- monitor.rs: Section monitoring logic with retry mechanisms (152 lines)
- enroll.rs: Enrollment logic with retry and notification (80 lines)
- state.rs: Application state management and health checks (177 lines)
- main.rs: Entry point and monitoring orchestration (278 lines)

All functionality preserved. Code builds successfully with no breaking changes.
